Move `Item` and `Store` groery components into components dir
Just some code cleanup; nothing to see here, folks! :)
