+++ app/javascript/groceries/components/item.vue
