+++ app/javascript/groceries/components/store.vue
