+++ app/javascript/groceries/groceries.vue
@@ -11,7 +11,7 @@
 import { mapState } from 'vuex';
 import LoggedInHeader from './components/logged_in_header.vue';
 import Sidebar from './components/sidebar.vue';
-import Store from './store.vue';
+import Store from './components/store.vue';
 
 export default {
   components: {
